handle unauthorize action

+++ app/Http/Controllers/PostsController.php
@@ -82,7 +82,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+
       $post = Post::find($id);
+      if(auth()->user()->id !== $post->user_id){
+        return redirect('/posts')->with('error', 'unauthorize page');
+      }
       return view('posts.edit')->with('post',$post);
     }
 
@@ -116,6 +120,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
           $post = Post::find($id);
+          if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'unauthorize page');
+          }
           $post->delete();
           return redirect('/posts')->with('success','post deleted');
     }
